extracao-dados-innovation/requirements.txt
opencv-python==4.5.5.64
pytesseract==0.3.9
Pillow==9.1.0
pdf2image==1.16.0
sklearn==0.0
scikit-learn==1.0.2
fuzzywuzzy==0.18.0
transformers==4.17.0
pyodbc==4.0.32
Flask==2.1.1
SQLAlchemy==1.4.34